\name{exams2openolat}
\alias{exams2openolat}

\title{Convenience Exams Interface for OpenOlat}

\description{
  Convenience interface to \code{\link[exams]{exams2openolat}} from the \pkg{exams}
  package with slightly changed defaults that also stores an RDS file with the
  xexams object returned.
}

\usage{
  exams2openolat(file, n = 1L, dir = ".", name = "olattest", adescription = "", ...)
}

\arguments{
  \item{file}{character. A specification of a (list of) exercise files.}
  \item{n}{integer. The number of copies to be compiled from \code{file}.}
  \item{dir}{character. The default is either display on the screen or
    the current working directory.}
  \item{name}{character. A name prefix for resulting exercises and RDS file.}
  \item{adescription}{character with assessment description.}
  \item{\dots}{arguments passed on to \code{\link[exams]{exams2openolat}}.}
}

\value{
  A list of exams as generated by \code{\link[exams]{xexams}} is
  returned invisibly.
}

\examples{
## load package and enforce par(ask = FALSE)
library("exams2openolat")
options(device.ask.default = FALSE)

## define an exams (= list of exercises)
myexam <- list(
  "boxplots",
  c("tstat", "ttest", "confint"),
  c("regression", "anova"),
  c("scatterplot", "boxhist"),
  "relfreq"
)

## output directory
mydir <- tempdir()

## generate .zip with OpenOlat exam in temporary directory
## using a few customization options
exams2openolat(myexam, n = 3, dir = mydir, maxattempts = 3)
dir(mydir)
}

\keyword{utilities}
