\name{olat_eval}
\alias{olat_eval}

\title{Evaluate OpenOlat Exams}

\description{
  Evaluate OpenOlat exams produced with \code{\link{exams2openolat}},
  and carried out and exported in OpenOlat.
}

\usage{
  olat_eval(file, plot = TRUE)
}

\arguments{
  \item{file}{character. Base file name of RDS and XLS file with
    exam generated by \code{\link{exams2openolat}} and exported
    OpenOlat results, respectively.}
  \item{plot}{logical. Should barplots with the aggregated
    results be displayed on the screen?}
}

\details{
  \code{olat_eval} is a companion function for \code{\link{exams2openolat}}.
  It evaluates the exams carried out in OpenOlat for further processing outside
  of OpenOlat (in CSV format).
}

\value{
  A \code{data.frame} with the detailed exam results is returned invisibly.
  It is also written to a CSV file in the current directory.
}

\seealso{\code{\link{exams2openolat}}}

\keyword{utilities}
